feat: add voice-over audio and checkpointing (#8)
# Overview

This PR adds voice-over narration to be merged into the video as an audio track.

Supported voice providers:
* OpenAI
* ElevenLabs

There is a very basic attempt to select a voice and rendering parameters to match the spirit of the ad.

## Checkpointing

A checkpointing feature is introduced, to allow the reuse of assets generated in past runs, to help reduce cost.